feat(common, physics): add ray_cast to ground
